requirements.txt
gradio>=4.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
groq>=0.4.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
#openai>=1.0.0
pandas>=2.0.0
pathlib>=1.0.0